requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2